ci: [UIP-6196]: replace & with && in astro build script
There was a file already exists `./lib` error in the astro build step. The build was passing but babel  wasn't processing the files.

This was caused by adding the line `& cp README.md ./lib & cp MIGRATION.md ./lib` to the end of the `build` script. This created the `./lib` directory. Therefor when babel ran and was also trying to create the same directory, it gave the error. Later when schema-form tried to access astro, it couldn't.

The solution was to replace `&` with `&&`, which forces the list of tasks to run in series instead of in parrelle, giving babel the change to finish before the README.md and MIGRATION.md files were copied over. 

## Describe your changes
1. Ran yarn locally to update yarn.lock, missing references to `@pingux/astro@^2.0.0-alpha.5`
2. Updated the astro build script to replace `&` with `&&`