fix: [UIP-6194] Update PageHeader Recipe to match mock changes
## Describe your changes
With Page Header component being deprecated, the warning says to replace the component with the code in the Page Header recipe, however that code is outdated and different from designs. The purpose of this MR is to ensure the page header recipe is in alignment with the current figma mock (not the future one).
https://www.figma.com/file/bpbEw54gTTcdIiZs4PKK1p/Astro-Specs?node-id=3408-15250&t=hVyBd95FjKiM9aLt-0

## Any testing tips

## Checklist before requesting a review
- [x] I have performed a self-review of my code
- [NA] I have added tests for any new or fixed functionality to prevent regressions
- [NA] I have added clarifying code comments to blocks which other devs may have questions on

Closes UIP-6194