fix: [UIP-6220] Down grade @emotion/react and @emotion/styled to 11.1.1
## Describe your changes
There is a package conflict with different versions of emotion, resulting in styles being off when upgrading from astro v1 to astro v2 in an external app. 

Removing rebass from the repo was part of the testing process, it didn't solve the issue but because we no longer need the package, I just left it removed.

## Any testing tips
Reproducing this bug requires a few steps
1. set up the external testing app
2. checkout commit `52214d9` and make a copy of the package.json and yarn.lock files.
3. go back to the original commit of you external app
4. past the contents of the copied files into there respective files
5. run `yarn` to get the package state to match that of the app with astro v1
6. return your package.json file to its original state representing astro v2, with 2.0.0-beta.0
7. run `yarn && yarn start`
8. this should result major styling errors across the external app. 

To observe the fix you then need to go through the previous steps 1-5, getting the package.json and yarn.lock file in a v1 state, then:
1. export this branch of astro using yalc
2. upgrade the external apps version of astro using yalc
3. run `yarn && yarn start`
4. This should result in a health version of the app with expected styles.

## Checklist before requesting a review
- [:white_check_mark: ] I have performed a self-review of my code
- [na] I have added tests for any new or fixed functionality to prevent regressions
- [na] I have added clarifying code comments to blocks which other devs may have questions on

Closes UIP-6220