fix: [UIP-6198] ImageUploadField Screen Reader Hidden Content
## Describe your changes
Add `display: "none"` to hidden input so screen reader ignores hidden content.
## Any testing tips
Can be a little finicky to reproduce. I use the VoiceOver in the Mac system settings. I find I am able to reproduce it by pressing `control` + `option`+ `shift` and using the arrow keys to navigate between the label and the input. In the screen recording, you can see at the end, it highlights a small square and announces a button group for a non-visible element. 


![Screen_Recording_2023-04-11_at_2.35.29_PM](/uploads/49b66da5ddb2ddd886d7fe7548e194c7/Screen_Recording_2023-04-11_at_2.35.29_PM.mov)
## Checklist before requesting a review
- [x] I have performed a self-review of my code
- [x] I have added tests for any new or fixed functionality to prevent regressions
- [x] I have added clarifying code comments to blocks which other devs may have questions on

Closes UIP-6198