chore: [UIP-6271] Image: add aria-busy, label and role attributes to skeleton
## Describe your changes
Added aria-label, role, and aria-busy attributes. Use background-image instead of background-color.

Responce from LA: 
```
I would suggest instead of using a standard background colour to represent the timed-out element, to instead use a CSS background image that produces the same visual output. You can then add an aria role and aria-label to the <div> to treat it as a functional image on the page. When positioned inside of the aria-live region, the screen reader would announce the updated content.
```

## Any testing tips

## Checklist before requesting a review
- [x] I have performed a self-review of my code
- [n/a] I have added tests for any new or fixed functionality to prevent regressions
- [n/a] I have added clarifying code comments to blocks which other devs may have questions on