chore: preparation to remove lerna from DockerFiles
## Describe your changes
1. Add `build-ci`, `coverage-ci` and `lint-ci` scripts to root package.json, that all use `yarn lerna`
2. Consume ^^ scripts in gitlab-ci and Jenkins files

The above will tell node to use the lerna version and files in the repo, rather than the globally provided version in the Docker container 

## Any testing tips
If the pipeline passes its working in gitlab.

## Note
This does not remove lerna from the docker files, this is just setup to make it possible. Once we can confirm that this worked, we can update the docker files.