fix: [UIP-6150] multiselect values use tab navigation instead of arrow keys
## Describe your changes
Under the hood end-user used react-select for it's Multiselect field. It's default behavior is to use arrow keys to be able to "focus" the selected values. It wasn't truly focusing the values, internally it was using state to only change the background-color. As of react-select@4.3.1, they do not expose an API to allow users to be able to control that state so we moved forward with not allowing users to use left and right arrow which controlled that state.

This PR includes: 
- I was not able to use react-selects <components.MutliValueRemove /> component and give it a tabIndex so I created my own and included the same CrossIcon for the 'X' button.
- the AriaLiveMessage.guidance under the switch case (value) updated to use Shift+Tab language instead of Left and Right Arrow. This may need to be reviewed for proper wording. the other cases are react-selects defaults.
- Prevent default behavior of being able to use ArrowLeft and ArrowRight keys.
- added prop backspaceRemovesValue to replicate astro's behavior. Wasn't sure about this.
## Any testing tips
I am having trouble getting Jest working. Wanted to get this up before adding some tests. For manual testing you can go to the [component](http://localhost:6006/?path=/story/components-multivalues--default) in Storybook and turn on Screen reader.

## Checklist before requesting a review
- [x] I have performed a self-review of my code
- [ ] I have added tests for any new or fixed functionality to prevent regressions
- [x] I have added clarifying code comments to blocks which other devs may have questions on

Closes UIP-6150