chore: [UIP-6308] Rocker Button: update hover & pressed state styles
The rocker button selected+hover should have 20% darker bg and selected+presed - 40%.
The 20% and 40% black should be mixed in with the base color, which can be customized.
The goal is to darken the color programmatically so that it works with whatever color they set the rocker button to.

## Describe your changes
- calculate darker colors

Figma: https://www.figma.com/file/bpbEw54gTTcdIiZs4PKK1p/Astro-Specs?type=design&node-id=7523-27245&t=AGNVw17GhCLQThZB-0

## Checklist before requesting a review
- [x] I have performed a self-review of my code
- [n/a] I have added tests for any new or fixed functionality to prevent regressions
- [n/a] I have added clarifying code comments to blocks which other devs may have questions on