fix: [UIP-6143] End User fix Anchor Tags in Preview
## Describe your changes
`href="#"` inside an iframe was causing the iframe to re-render inside itself. Removing the href creates a placeholder tag that links nowhere, it this preview state, it feels appropriate. A little research was done to verify this doesn't break accessibly but I didn't dig real deep.
 
## Any testing tips
Click all the links in Display=>Preview=>Default to make sure the are all disabled.

## Checklist before requesting a review
- [:heavy_check_mark: ] I have performed a self-review of my code
- [na] I have added tests for any new or fixed functionality to prevent regressions
- [na] I have added clarifying code comments to blocks which other devs may have questions on

Closes UIP-6143