fix: [UIIP-6564] HelpHint Tooltip Screenreader
## Describe your changes
Added `role='status'` so screenreader will read content of popover. Adding this role will add `aria-live="polite"` by default.

Updated useEffect hook to keep the popover visible while the trigger has focus. According to [docs](https://www.w3.org/WAI/WCAG21/Understanding/content-on-hover-or-focus.html) this is best practice for components like this:

"The intent of this condition is to ensure users have adequate time to perceive the additional content after it becomes visible. Users with disabilities may require more time for many reasons, such as to change magnification, move the pointer, or simply to bring the new content into their visual field. Once it appears, the content should remain visible until:

The user removes hover or focus from the trigger and the additional content, consistent with the typical user experience;
The user dismisses the additional content via the mechanism provided to satisfy the Dismissable condition; or
The information conveyed by the additional content becomes invalid, such as a 'busy' message that is no longer valid." 
## Any testing tips

## Checklist before requesting a review
- [ ] I have performed a self-review of my code
- [ ] I have added tests for any new or fixed functionality to prevent regressions
- [ ] I have added clarifying code comments to blocks which other devs may have questions on

Closes UIP-6564