feat: [UIP-6363]: create panel content recipe
## Describe your changes
The intent of this ticket was to create custom variants to help with the styling of Panel Content but my observation after working with it for a while is that if the the Panel Content recipe is built with best practices in mind, and some of the new Astro features are applied it can be done mostly without custom styles. The few places where custom variants could be applied they feel to be either low value or so specific they would be confusing and/or hard to use.

Using the new PanelHeader, ButtonBar, EditButton and added Accordion functionality, removes a lot of code and styling. Also focusing on using variants and tools that are currently in the library makes for a good demonstration of how to best build out Panel Content.

## Note:
The Edit Panel spec shows the PanelHeader with a Breadcrumb, this is not implemented in this MR at the moment. There is an [MR that adds this functionality](https://pingidentity.atlassian.net/browse/UIP-6605) to PanelHeader and so it will be added to the Edit Panel as a follow up.

## Checklist before requesting a review
- [:checkered_flag: ] I have performed a self-review of my code
- [:checkered_flag:  ] I have added tests for any new or fixed functionality to prevent regressions
- [:checkered_flag:  ] I have added clarifying code comments to blocks which other devs may have questions on

Closes UIP-6363