refactor: [UIP-6677]: add universalComponentTest
## Describe your changes
This is one of two MR's that will close UIP-6677.

This one adds `universalComponentTest`, that will replace the `axeTest`. `universalComponentTest` expands the `axeTest` by adding the `should forward refs properly` test. 

A few of the test files have be updated with this change to verify that it works, but the rest will be done in a follow up MR, that will close UIP-6677.

## Any testing tips
Verify that if ref forwarding is removed from any of the updated test, `should forward refs properly` should fail.

## Checklist before requesting a review
- [:white_check_mark: ] I have performed a self-review of my code
- [:white_check_mark: ] I have added tests for any new or fixed functionality to prevent regressions
- [:white_check_mark:  ] I have added clarifying code comments to blocks which other devs may have questions on