test: [UIP-6707]: AstroProvider: replace axeTest with universalComponentTests
## Describe your changes
AstroProvider was providing the forwarded ref to the `ThemeProvider`. This was failing the forward ref test and didn't make sense to me so I moved the ref to the Box wrapper inside of AstroProvider. 

## Any testing tips
No UI changes expected

I'm concerned that moving the ref around might be a breaking change. I'm open to feedback and discussion around this.

## Checklist before requesting a review
- [ check ] I have performed a self-review of my code
- [ check ] I have added tests for any new or fixed functionality to prevent regressions
- [ NA ] I have added clarifying code comments to blocks which other devs may have questions on

Closes UIP-6707