feat: [UIP-6840] PhoneInputField improvements
## Describe your changes
The country name string has been removed. The output (e.target.name) gives the country name with the country extension. United States is the first option in the list.The value obtained from the widget is concatenated with the country code. Selecting the same country from combobox gives consistent result. A controlled filter has been added to search country name as well as country extension and country-list has been updated to latest.

## Any testing tips
To test if there are large spacing in between elements, add the code snippet below to the ControlledMenu Story of combobox and resize the size of the window to multiple width to make sure that there are no large spaces between the list elements. For a small time the popover menu shifts to the left before the correct positioning if anyone has has any idea why please let me know.

```
import { countries as countriesObj } from 'countries-list';
export const ControlledMenu = () => {
  const [isOpen, setIsOpen] = useState(false);

  // Use menuTrigger="manual" if you don't want open change events to fire on input / focus
  return (
    <OverlayProvider>
    <Box 
    sx={{
        width: '100%',
        maxWidth: '500px',
        position: 'relative',
        minWidth:'200px',
      }}
      >
    <ComboBoxField
      label="Example label"
      defaultItems={Object.entries(countriesObj)}
      {...actions}
      isOpen={isOpen}
      onOpenChange={setIsOpen}
      width="100%"
      customWidth={isOpen && '200px'}
      wrapperProps={{
        sx: {
          position: 'absolute',
          transition: '0.2s width ease',
          width: isOpen  ? '100%' : '110px',
         
        },
      }}
    >
      {item => (
        <Item key={item[0]}>
          {`${item[1].name}${item[1].name !== item[1].native ? ` (${item[1].native})` : ''} +${item[1].phone.split(',')[0]}`}
        </Item>
      )}
    </ComboBoxField>
    </Box>
  </OverlayProvider>
  );
};

```

## Checklist before requesting a review
- [ ] I have performed a self-review of my code
- [ ] I have added tests for any new or fixed functionality to prevent regressions
- [ ] I have added clarifying code comments to blocks which other devs may have questions on

Closes UIP-6840