feat: [UIP-7157] CheckboxField fix selecting
## Describe your changes
Revert change from a previous MR that caused regression. That change was unwrapping the checkbox in the label. The Checkbox is now wrapped in the label. Opacity can never be higher than its parent so I made checkbox labels have an opacity of 1 when is disabled and then set the alpha value of the label itself to .5 so that the label text still has opacity while the checkbox maintains 100% opacity.
## Any testing tips
- Confirm clicking label or checkbox toggles the checkbox.
- Confirm styling matches figma: https://www.figma.com/design/Oa6VYtJcUJzEJuuRp0p4ls/Astro?node-id=6938-65&node-type=frame&t=v5gP9V7ax2WkLyrP-0 Especially for the disabled stories. The modifying disabled color ticket is what originally caused the regression.

## Checklist before requesting a review
- [ ] I have performed a self-review of my code
- [ ] I have added tests for any new or fixed functionality to prevent regressions
- [ ] I have added clarifying code comments to blocks which other devs may have questions on

Closes UIP-7157